requirements.txt
piper-tts
pygame
faster-whisper
sounddevice
numpy
pvporcupine
pyaudio
python-dotenv
openai
ddgs
wikipedia